Move ProgressBar.jsx into different directory
